refactored out createDeployment
