+++ packages/wrangler/src/pages/build.tsx
@@ -9,7 +9,7 @@ import { buildWorker } from "./functions/buildWorker";
 import { generateConfigFromFileTree } from "./functions/filepath-routing";
 import { writeRoutesModule } from "./functions/routes";
 import { pagesBetaWarning, RUNNING_BUILDERS } from "./utils";
-import type { Config} from "./functions/routes";
+import type { Config } from "./functions/routes";
 import type { ArgumentsCamelCase, Argv } from "yargs";
 
 type PagesBuildArgs = {
