+++ packages/wrangler/src/pages/functions/buildPlugin.ts
@@ -24,7 +24,9 @@ export function buildPlugin({
   onEnd = () => {},
 }: Options) {
   return build({
-    entryPoints: [resolve(__dirname, "../src/pages/templates/template-plugin.ts")],
+    entryPoints: [
+      resolve(__dirname, "../src/pages/templates/template-plugin.ts"),
+    ],
     inject: [routesModule],
     bundle: true,
     format: "esm",
